+++ plugins/soul-backend-sample/src/soul_backend_sample/__init__.py
@@ -1,4 +1,4 @@
-﻿# SPDX-FileCopyrightText: 2024-present Codex Soul System Contributors
+# SPDX-FileCopyrightText: 2024-present Codex Soul System Contributors
 # SPDX-License-Identifier: MIT
 
 from ._plugin import __plugin_interface_version__, register_backends, JsonBackend