+++ soul/__init__.py
@@ -1,4 +1,4 @@
-﻿from .__about__ import __version__, __plugin_interface_version__, SOUL_SCHEMA_VERSION
+from .__about__ import __version__, __plugin_interface_version__, SOUL_SCHEMA_VERSION
 from ._soul import Soul
 from ._base import SoulBackend
 from ._types import (