+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿# Changelog
+# Changelog
 
 ## v1.4.0 (2026-06-04)